backend/requirements.txt
channels==3.0.4
channels-redis==3.3.1
django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
requests==2.26.0